requirements.txt
Pillow
ImageHash
pdf2image